requirements.txt
streamlit
python-pptx~=0.6.21
PyPDF2
io
spacy